[Refactor] Factors out initialisation functions in different files
